Validate token in dashboard group folder
